backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
